+++ libs/blocks/src/lib/live-market/live-market.spec.tsx
@@ -7,13 +7,12 @@ import LiveMarket from '.';
 describe('LiveMarket', () => {
   beforeEach(() => {
     render(
-      <LiveMarket
+      <LiveMarket.Slider
         link={{
           href: '/',
           content: 'More forex',
         }}
         header="header here"
-        isHeaderSection={true}
         disclaimer="Disclaimer here"
       />,
       {